Moved functionality to public, as there is no import from server that needs the function on common
