+++ common/helpers.ts
@@ -11,30 +11,6 @@ import { Get, Set } from '../../../src/plugins/opensearch_dashboards_utils/commo
 
 // Wazuh
 import { AllowedActionsBySpace, SpaceTypes, UserSpacesOrder } from './constants';
-import { getCapabilities } from '../public/services/utils/constants';
-
-export const UI_DISABLED_SETTINGS_IDS = {
-  INDEX_DISCARDED_EVENTS: 'index-discarded-events',
-  INDEX_UNCLASSIFIED_EVENTS: 'index-unclassified-events',
-  INDEX_RAW_EVENTS: 'index-raw-events',
-} as const;
-
-export type UIDisabledSettingId =
-  (typeof UI_DISABLED_SETTINGS_IDS)[keyof typeof UI_DISABLED_SETTINGS_IDS];
-
-export const isUiSettingDisabled = (settingId: UIDisabledSettingId): boolean => {
-  const caps = getCapabilities()?.securityAnalytics;
-
-  if (!caps) return false;
-
-  const capabilityMap: Record<UIDisabledSettingId, string> = {
-    [UI_DISABLED_SETTINGS_IDS.INDEX_DISCARDED_EVENTS]: 'showIndexDiscardedEvents',
-    [UI_DISABLED_SETTINGS_IDS.INDEX_UNCLASSIFIED_EVENTS]: 'showIndexUnclassifiedEvents',
-    [UI_DISABLED_SETTINGS_IDS.INDEX_RAW_EVENTS]: 'showIndexRawEvents',
-  };
-
-  return caps[capabilityMap[settingId]] === false;
-};
 
 export function aggregateMetrics(
   metrics: PartialMetricsCounter,
